+++ index.html
@@ -10,7 +10,7 @@
 </head>
 
 <body>
-  <script src="js/main.js"></script>
+
   <main>
 
     <!-- Фильтрация изображений от других пользователей -->
@@ -235,5 +235,6 @@
       </section>
     </template>
 
+    <script src="js/main.js" type="module"></script>
 </body>
 </html>
